Separate columns data
